requirements.txt
asgiref==3.8.1
black==24.8.0
build==1.2.2.post1
cachecontrol==0.14.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.1
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
exceptiongroup==1.2.2
fastjsonschema==2.20.0
filelock==3.16.1
flake8==7.1.1
gunicorn==23.0.0
h11==0.14.0
identify==2.6.1
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
keyring==24.3.1
mccabe==0.7.0
more-itertools==10.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pre-commit==4.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyproject-hooks==1.2.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
rapidfuzz==3.10.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
shellingham==1.5.4
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
tomlkit==0.13.2
trove-classifiers==2024.9.12
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
virtualenv==20.26.6
whitenoise==6.7.0
xattr==1.1.0
zipp==3.20.2
